refactor: Players 클래스 패키지 변경
- /repository -> /domain/user
