+++ src/main/java/domain/user/Players.java
@@ -1,6 +1,5 @@
-package repository;
+package domain.user;
 
-import domain.user.Player;
 import java.util.ArrayList;
 import java.util.List;
